+++ demo/components/Application.js
@@ -1,61 +1,34 @@
-import PanelGroup from 'react-panelgroup';
-import React      from 'react';
-import {getSchemas} from '/assets/schemas';
-import {getSchema}  from '/assets/schemas';
-import {getStyles}  from '/assets/styles';
-import {getThemes}  from '/assets/themes';
-import {getTheme}   from '/assets/themes';
+import React       from 'react';
+import {Component} from 'react';
 
-export class Application extends React.Component {
+import presets from '../lib/presets';
+import schema  from '../lib/schema';
+
+import ApplicationForm         from './ApplicationForm';
+import ApplicationPreviewField from './ApplicationPreviewField';
+import ApplicationPropsField   from './ApplicationPropsField';
+import ApplicationSelectField  from './ApplicationSelectField';
+
+class Application extends Component {
     constructor () {
         super(...arguments);
 
-        this.state = {
-            doc: null,
-            showInlineError: false,
-            schemas: getSchemas(),
-            themes:  getThemes(),
+        this.state = schema.clean({});
 
-            schema: getSchema(),
-            styles: getStyles(),
-            theme:  getTheme()
-        };
-
-        this.onSchema = this.onSchema.bind(this);
-        this.onTheme  = this.onTheme.bind(this);
-        this.changeInlineError  = this.changeInlineError.bind(this);
+        this.onChange = this.onChange.bind(this);
     }
 
-    onSchema ({target: {value}}) {
-        this.setState({doc: null, schema: getSchema(value)});
-    }
+    onChange (key, value) {
+        if (key === 'preset') {
+            this.setState({props: {...this.state.props, schema: presets[value]}});
+        }
 
-    onTheme ({target: {value}}) {
-        this.setState({doc: null, styles: getStyles(value), theme: getTheme(value)});
-    }
-    changeInlineError () {
-        this.setState({showInlineError: !this.state.showInlineError});
+        this.setState({[key]: value});
     }
 
     render () {
-        const {
-            state: {
-                doc,
-                schema,
-                schemas,
-                styles,
-                theme,
-                themes,
-                showInlineError
-            },
-
-            onSchema,
-            onTheme,
-            changeInlineError
-        } = this;
         return (
-            <PanelGroup spacing={3} >
-
+            <ApplicationForm label={false} model={this.state} onChange={this.onChange} schema={schema} spacing={3}>
                 <section>
                     <nav className="panelGroupNavbar">
                         <section className="panelHeader">
@@ -65,70 +38,26 @@ export class Application extends React.Component {
                         </section>
 
                         <section className="panelHeaderInfo">
-                            <select value={theme.text} onChange={onTheme}>
-                                {themes.map(theme =>
-                                    <option key={theme} value={theme}>
-                                        Theme - {theme}
-                                    </option>
-                                )}
-                            </select>
-
-                            <select value={0} onChange={onSchema}>
-                                <option value={0}>
-                                    Select example
-                                </option>
-
-                                {schemas.map(schema =>
-                                    <option key={schema} value={schema}>
-                                        {schema}
-                                    </option>
-                                )}
-                            </select>
-                            <input type="checkbox" onChange={changeInlineError} />
+                            <ApplicationSelectField name="preset" />
+                            <ApplicationSelectField name="theme" />
                         </section>
                     </nav>
 
-                    <textarea
-                        spellCheck={false}
-                        value={schema.string}
-                        onChange={onSchema}
-                        style={{fontSize: '13px', lineHeight: '1'}}
-                    />
+                    <ApplicationPropsField name="props" spellCheck={false} />
+
+                    <a href="https://github.com/vazco/uniforms">
+                        {/* eslint-disable max-len */}
+                        <img
+                            alt="Fork me on GitHub"
+                            data-canonical-src="https://s3.amazonaws.com/github/ribbons/forkme_right_darkblue_121621.png"
+                            src="https://camo.githubusercontent.com/38ef81f8aca64bb9a64448d0d70f1308ef5341ab/68747470733a2f2f73332e616d617a6f6e6177732e636f6d2f6769746875622f726962626f6e732f666f726b6d655f72696768745f6461726b626c75655f3132313632312e706e67"
+                        />
+                        {/* eslint-enable max-len */}
+                    </a>
                 </section>
-                <section style={{margin: '20px', fontSize: '13px', lineHeight: '1', overflow: 'scroll'}}>
-                    {styles}
-                    <div>
-                        {schema.object ? (
-                            <theme.components.AutoForm
-                                key={schema.string}
-                                schema={schema.object}
-                                showInlineError={showInlineError}
-                                fieldData={{'colors' : ['#cc0000', '#000099', '#808080', '#da4264']}}
-                                onSubmit={doc => {
-                                    console.log(doc);
-                                    this.setState({doc: JSON.stringify(doc, null, 4)});
-                                }}
-                            />
-                        ) : (
-                            <span>
-                                {schema.error.message}
-                            </span>
-                        )}
-                    </div>
-                    <div>
-                        {!!doc && (
-                            <br />
-                        )}
-                    </div>
-                    <div>
-                        {!!doc && (
-                            <pre>
-                                {doc}
-                            </pre>
-                        )}
-                    </div>
-                </section>
-            </PanelGroup>
+
+                <ApplicationPreviewField name="props" nameTheme="theme" />
+            </ApplicationForm>
         );
     }
 }