+++ demo/lib/styles.js
@@ -1,0 +1,24 @@
+import React from 'react';
+
+const style = links => links.map((link, index) => <link key={index} rel="stylesheet" href={link} />);
+const styles = {
+    antd: style([
+        'https://cdnjs.cloudflare.com/ajax/libs/antd/2.3.2/antd.min.css'
+    ]),
+    bootstrap3: style([
+        'https://cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/3.3.6/css/bootstrap.min.css'
+    ]),
+
+    bootstrap4: style([
+        'https://cdnjs.cloudflare.com/ajax/libs/octicons/3.5.0/octicons.min.css',
+        'https://cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/4.0.0-alpha.5/css/bootstrap.min.css'
+    ]),
+
+    semantic: style([
+        'https://cdnjs.cloudflare.com/ajax/libs/semantic-ui/2.2.4/semantic.min.css'
+    ]),
+
+    unstyled: style([])
+};
+
+export default styles;