+++ packages/uniforms-unstyled/src/NestField.js
@@ -1,0 +1,33 @@
+import React          from 'react';
+import connectField   from 'uniforms/connectField';
+import filterDOMProps from 'uniforms/filterDOMProps';
+import injectName     from 'uniforms/injectName';
+import joinName       from 'uniforms/joinName';
+
+import AutoField from './AutoField';
+
+const Nest = ({
+    children,
+    fields,
+    label,
+    name,
+    ...props
+}) =>
+    <section {...filterDOMProps(props)}>
+        {label && (
+            <label>
+                {label}
+            </label>
+        )}
+
+        {children ? (
+            injectName(name, children)
+        ) : (
+            fields.map(key =>
+                <AutoField key={key} name={joinName(name, key)} />
+            )
+        )}
+    </section>
+;
+
+export default connectField(Nest, {includeInChain: false});