+++ packages/uniforms-unstyled/src/components/forms/BaseForm.js
@@ -1,9 +1,0 @@
-import {BaseForm} from 'uniforms';
-
-const Unstyled = parent => class extends parent {
-    static Unstyled = Unstyled;
-
-    static displayName = `Unstyled${parent.displayName}`;
-};
-
-export default Unstyled(BaseForm);